Move `-l` option parsing into its own submodule
No functional change (yet).
